// wealth-mgmt/src/app/models/investment.ts
export interface Investment {

  investmentId?: number;

  portfolioId?: number;

  portfolioName?: string;

  investmentName?: string;

  investmentType?: string;

  quantity?: number;

  purchasePrice?: number;

  currentPrice?: number;

  totalValue?: number;

  purchaseDate?: string;

  status?: string;

}